requirements.txt
crewai>=0.28.0
groq>=0.4.0
fastapi>=0.104.0
uvicorn>=0.23.2
jinja2>=3.1.2
pydantic>=2.4.2
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
aiofiles>=23.2.0